Demo/Improper Parking/requirements.txt
flask==3.1.0
google-generativeai==0.8.4
pillow==11.1.0
python-dotenv==1.0.1 